+++ src/api/flow/dto/index.ts
@@ -1,1 +1,2 @@
+export * from './create-flow.dto';
 export * from './created-flow.dto';
